+++ Bootstrap/Utils/Console.h
@@ -10,6 +10,7 @@ public:
 	static bool GeneratingAssembly;
 	static bool AlwaysOnTop;
 	static bool HideWarnings;
+	static bool UseManualColoring;
 	enum DisplayMode
 	{
 		NORMAL,